+++ chado/modules/map/map.sql
@@ -15,11 +15,11 @@
 -- ================================================
 
 create table featuremap (
-    featuremap_id serial not null,
+    featuremap_id bigserial not null,
     primary key (featuremap_id),
     name varchar(255),
     description text,
-    unittype_id int null,
+    unittype_id bigint null,
     foreign key (unittype_id) references cvterm (cvterm_id) on delete set null INITIALLY DEFERRED,
     constraint featuremap_c1 unique (name)
 );
@@ -29,19 +29,19 @@ create table featuremap (
 -- ================================================
 
 create table featurerange (
-    featurerange_id serial not null,
+    featurerange_id bigserial not null,
     primary key (featurerange_id),
-    featuremap_id int not null,
+    featuremap_id bigint not null,
     foreign key (featuremap_id) references featuremap (featuremap_id) on delete cascade INITIALLY DEFERRED,
-    feature_id int not null,
+    feature_id bigint not null,
     foreign key (feature_id) references feature (feature_id) on delete cascade INITIALLY DEFERRED,
-    leftstartf_id int not null,
+    leftstartf_id bigint not null,
     foreign key (leftstartf_id) references feature (feature_id) on delete cascade INITIALLY DEFERRED,
-    leftendf_id int,
+    leftendf_id bigint,
     foreign key (leftendf_id) references feature (feature_id) on delete set null INITIALLY DEFERRED,
-    rightstartf_id int,
+    rightstartf_id bigint,
     foreign key (rightstartf_id) references feature (feature_id) on delete set null INITIALLY DEFERRED,
-    rightendf_id int not null,
+    rightendf_id bigint not null,
     foreign key (rightendf_id) references feature (feature_id) on delete cascade INITIALLY DEFERRED,
     rangestr varchar(255)
 );
@@ -61,13 +61,13 @@ COMMENT ON COLUMN featurerange.featuremap_id IS 'featuremap_id is the id of the 
 -- ================================================
 
 create table featurepos (
-    featurepos_id serial not null,
+    featurepos_id bigserial not null,
     primary key (featurepos_id),
-    featuremap_id serial not null,
+    featuremap_id bigserial not null,
     foreign key (featuremap_id) references featuremap (featuremap_id) on delete cascade INITIALLY DEFERRED,
-    feature_id int not null,
+    feature_id bigint not null,
     foreign key (feature_id) references feature (feature_id) on delete cascade INITIALLY DEFERRED,
-    map_feature_id int not null,
+    map_feature_id bigint not null,
     foreign key (map_feature_id) references feature (feature_id) on delete cascade INITIALLY DEFERRED,
     mappos float not null
 );
@@ -84,11 +84,11 @@ links to the feature (map) upon which the feature is being localized.';
 -- ================================================
 
 create table featuremap_pub (
-    featuremap_pub_id serial not null,
+    featuremap_pub_id bigserial not null,
     primary key (featuremap_pub_id),
-    featuremap_id int not null,
+    featuremap_id bigint not null,
     foreign key (featuremap_id) references featuremap (featuremap_id) on delete cascade INITIALLY DEFERRED,
-    pub_id int not null,
+    pub_id bigint not null,
     foreign key (pub_id) references pub (pub_id) on delete cascade INITIALLY DEFERRED
 );
 create index featuremap_pub_idx1 on featuremap_pub (featuremap_id);
